rubric: consolidate cli reference types under _shared/FunctionTypes
The shared YAML reference / indexed-check / resolve-outcome types lived
in cli/types.ts (YamlReferenceYml) and cli/utils/rubricReferences.ts
(YamlReference, IndexedCheck, ResolveOutcome). Each was the canonical
declaration for its file, but they're cross-cutting types that any
edge function dealing with rubric YAML may want to import. Per the
codebase convention they belong in supabase/functions/_shared/.

YamlReferenceYml is kept as a legacy alias so existing call sites
continue to compile while we converge on the canonical YamlReference
name. No runtime behavior changes; the type structures are identical.
