+++ supabase/functions/_shared/FunctionTypes.d.ts
@@ -488,3 +488,29 @@ export type CreateInvitationResponse = {
     error: string;
   }>;
 };
+
+// ─── Rubric reference types (shared between cli/ and any Edge Function that
+// imports/exports rubric YAML). Mirrors `YamlReference` in
+// `utils/supabase/DatabaseTypes.d.ts`; kept duplicate-but-aligned because the
+// frontend runs on Node and Edge Functions run on Deno.
+
+export type YamlReference = {
+  review_round?: string;
+  part?: string;
+  criterion?: string;
+  check?: string;
+  id?: number;
+};
+
+export type IndexedCheck = {
+  checkId: number;
+  checkName: string;
+  partName: string;
+  criterionName: string;
+  reviewRound: string | null;
+  rubricId: number;
+  assignmentId: number;
+  classId: number;
+};
+
+export type ResolveOutcome = { ok: true; target: IndexedCheck } | { ok: false; reason: string };
