+++ supabase/functions/cli/types.ts
@@ -32,18 +32,14 @@ export interface RubricWithHierarchy extends RubricRow {
 
 // ─── Rubric import/export structures ───────────────────────────────────────
 
+import type { YamlReference } from "../_shared/FunctionTypes.d.ts";
+
 /**
- * YAML reference entry on a rubric check, either name-keyed
- * (review_round + part + criterion + check) or by numeric `id` fallback.
- * Mirrors `YamlReference` in `utils/supabase/DatabaseTypes.d.ts`.
+ * YAML reference entry on a rubric check. Re-exported here under the legacy
+ * `YamlReferenceYml` name; the canonical declaration lives in
+ * `_shared/FunctionTypes.d.ts` so any Edge Function can use it.
  */
-export interface YamlReferenceYml {
-  review_round?: string;
-  part?: string;
-  criterion?: string;
-  check?: string;
-  id?: number;
-}
+export type YamlReferenceYml = YamlReference;
 
 export interface RubricExportCheck {
   name: string;
